Variables de entorno y modos

+++ src/index.ts
@@ -34,7 +34,7 @@ class Server {
     // No arrancamos hasta qye MariaDB esté lista
     await mariaDB.start();
     // Si queremos tirar la base de datos y comenzar desde cero, si no solo poner sync(), sin force
-    if (process.env.NODE_ENV !== 'test') {
+    if (process.env.NODE_ENV !== 'test' && env.DB_SYNC) {
       await mariaDB.getConnection().sync({ force: true });
       console.log(chalk.yellow('🗃  Tablas borradas y re-sincronizadas'));
     }
